+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.31.0] - 2025-02-12
+
+### Added
+
+- Added `list_objects` method to storage clients to list available objects for a given datapoint.
+- Added `download_objects` method to storage clients to download a subset of the available objects for a given datapoint.
+
 ## [0.30.2] - 2025-01-27
 
 ### Changed
@@ -104,7 +111,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Released packages: `tilebox-datasets`, `tilebox-workflows`, `tilebox-storage`, `tilebox-grpc`
 
 
-[Unreleased]: https://github.com/tilebox/tilebox-python/compare/v0.30.2...HEAD
+[Unreleased]: https://github.com/tilebox/tilebox-python/compare/v0.31.0...HEAD
+[0.31.0]: https://github.com/tilebox/tilebox-python/compare/v0.30.2...v0.31.0
 [0.30.2]: https://github.com/tilebox/tilebox-python/compare/v0.30.1...v0.30.2
 [0.30.1]: https://github.com/tilebox/tilebox-python/compare/v0.30.0...v0.30.1
 [0.30.0]: https://github.com/tilebox/tilebox-python/compare/v0.29.0...v0.30.0